chore(common-utils): Split tests into matching files
In 6d34ce5 the implementations were split across multiple files, but the
tests were forgotten. Make good for that now.
